04_Agent_with_RAG/pyproject.toml
[project]
name = "04-agent-with-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
]
